worker/requirements.txt
influxdb-client==1.26.0
kafka-python==2.0.2
